backend/requirements.txt
amqp==2.1.4
appdirs==1.4.3
billiard==3.5.0.2
celery==4.0.2
Django==1.11.1
django-cors-headers==2.0.2
django-suit==0.2.25
djangorestframework==3.6.3
kombu==4.0.2
packaging==16.8
psycopg2==2.7.1
pyparsing==2.2.0
pytz==2017.2
six==1.10.0
vine==1.1.3
